feat: split types into multiple files
