+++ src/types/addressbook.ts
@@ -1,0 +1,5 @@
+
+export type Contact = {
+  email: string
+  name?: string
+}
