+++ src/types/options.ts
@@ -1,70 +1,13 @@
-import type { IcsCalendar, IcsEvent, IcsRecurrenceId } from 'ts-ics'
-import type { DAVCalendar } from 'tsdav'
-
-export type DomEvent = GlobalEventHandlersEventMap[keyof GlobalEventHandlersEventMap]
+import type { IcsEvent } from 'ts-ics'
+import type { Calendar, CalendarEvent } from './calendar'
+import type { Contact } from './addressbook'
+import type { attendeeRoleTypes, availableViews } from '../contants'
 
 export type RecursivePartial<T> = {
   [P in keyof T]?: RecursivePartial<T[P]>
 }
 
-export type Contact = {
-  email: string
-  name?: string
-}
-
-// TODO - CJ - 2025-07-03 - add <TCalendarUid = any> generic
-// TODO - CJ - 2025-07-03 - add options to support IcsEvent custom props
-export type Calendar = DAVCalendar & {
-  // INFO - CJ - 2025-07-03 - Useful fields from 'DAVCalendar'
-  // ctag?: string
-  // description?: string;
-  // displayName?: string | Record<string, unknown>;
-  // calendarColor?: string
-  // url: string
-  // fetchOptions?: RequestInit
-  headers?: Record<string, string>
-  uid?: unknown
-}
-
-export type CalendarObject = {
-  data: IcsCalendar
-  etag?: string
-  url: string
-  calendarUrl: string
-}
-
-export type EventUid = {
-  uid: string
-  recurrenceId?: IcsRecurrenceId
-}
-
-export const attendeeRoleTypes = [
-  'CHAIR',
-  'REQ-PARTICIPANT',
-  'OPT-PARTICIPANT',
-  'NON-PARTICIPANT',
-] as const
-export type IcsAttendeeRoleType = typeof attendeeRoleTypes[number]
-
-export const namedRRules = [
-  'FREQ=DAILY',
-  'FREQ=WEEKLY',
-  'BYDAY=MO,TU,WE,TH,FR;FREQ=DAILY',
-  'INTERVAL=2;FREQ=WEEKLY',
-  'FREQ=MONTHLY',
-  'FREQ=YEARLY',
-] as const
-
-export const availableViews = [
-  'timeGridDay',
-  'timeGridWeek',
-  'dayGridMonth',
-  'listDay',
-  'listWeek',
-  'listMonth',
-  'listYear',
-] as const
-export type View = typeof availableViews[number]
+export type DomEvent = GlobalEventHandlersEventMap[keyof GlobalEventHandlersEventMap]
 
 export type ServerSource = {
   serverUrl: string
@@ -83,6 +26,10 @@ export type AddressBookFn = {
   fetchContacts: () => Promise<Contact[]>
 }
 
+export type View = typeof availableViews[number]
+export type IcsAttendeeRoleType = typeof attendeeRoleTypes[number]
+
+
 export type SelectedCalendar = {
   url: string
   selected: boolean
@@ -99,15 +46,6 @@ export type SelectCalendarHandlers = {
   onClickSelectCalendars: (info: SelectCalendarsClickInfo) => void,
 }
 
-export type CalendarEvent = {
-  calendarUrl: string
-  event: IcsEvent
-}
-export type DisplayedCalendarEvent = {
-  calendarUrl: string
-  event: IcsEvent
-  recurringEvent?: IcsEvent
-}
 
 export type EventBodyInfo = {
   calendar: Calendar
